signer: move HSMSigner to _hsm_signer.py
This allows to hide all the import case-handling for optional
dependencies from the _signer module.
